WIP separate decoder and encoder
