+++ pkcs7.go
@@ -16,6 +16,7 @@ import (
 	"errors"
 	"fmt"
 	"hash"
+	"io"
 	"math/big"
 	"sort"
 	"time"
@@ -535,6 +536,7 @@ func (p7 *PKCS7) UnmarshalSignedAttribute(attributeType asn1.ObjectIdentifier, o
 
 // SignedData is an opaque data structure for creating signed data payloads
 type SignedData struct {
+	w             io.Writer
 	sd            signedData
 	certs         []*x509.Certificate
 	messageDigest []byte
